Bug 1184283 - Widget vsync source/observer From 56797f8b982b340151770938af74f8b6cef1ed14 Mon Sep 17 00:00:00 2001 Adds add/remove vsync observer to nsIWidget; Implement a VsyncSource in nsBaseWidget
There is an internal VsyncForwardingObserver in nsBaseWidget that is
both a VsyncSource and a VsyncObserver.  It observes whatever the widget
is currently listening to (a parent widget, a direct source id which
might be remote, etc.) and forwards it to anything that is related to
that widget (Compositors, RefreshDrivers, other child widgets).

We track widget parent changes so that we can listen to the new parent
appropriately.